refactor: move mount probing logic and models from preflight to disks
- MountToolCapabilities, MountCapabilities -> disks/models.py
- _probe_mount_tools, _check_mount_capabilities, check_mount_status
  -> disks/mount_checks.py
- Host-probing query primitives -> remote/queries.py (preflight/queries.py
  becomes a re-export shim)

This eliminates the disks -> preflight module dependency.
